requirements.txt
sqlalchemy
gino
asyncpg
aiogram==2.11.2

python-dotenv

